refactor #144: 함께 배달 DTO 이관 및 Valid추가
