+++ src/main/java/fastcampus/team7/Livable_officener/repository/DeliveryRepositoryCustom.java
@@ -1,10 +1,10 @@
 package fastcampus.team7.Livable_officener.repository;
 
-import fastcampus.team7.Livable_officener.dto.RoomDetailDTO;
+import fastcampus.team7.Livable_officener.dto.delivery.RoomDetailDTO;
 
 import java.util.List;
 
-import static fastcampus.team7.Livable_officener.dto.ChatRoomListResponseDTO.ChatRoomListDTO;
+import static fastcampus.team7.Livable_officener.dto.delivery.ChatRoomListResponseDTO.ChatRoomListDTO;
 
 public interface DeliveryRepositoryCustom {
     RoomDetailDTO findRoomById(Long id);
